:art: move presentation layer to component
